// edge-event-hotspot-relay/src/event-relay-mcp/raindrop.gen.ts
// Code generated by raindrop build, DO NOT EDIT.
import { Ai, Annotation, App, KvCache, Logger, MRNObject, SmartBucket, SmartSql, Tracer,  } from '@liquidmetal-ai/raindrop-framework';

export interface Env {
  _raindrop: {
    app: App;
  };
  AI: Ai;
  ALERTS: SmartBucket;
  annotation: Annotation<Omit<MRNObject, 'type' | 'applicationName' | 'versionId'>>;
  EDGE_EVENTS: SmartBucket;
  HOTSPOT_DB: SmartSql;
  logger: Logger;
  mem: KvCache;
  tracer: Tracer;
}